requirements.txt
requests
pandas
beautifulsoup4
yfinance
jinja2
PyYAML
python-dateutil
gspread
google-auth
fredapi
python-dotenv
